Changd InputGood's Quantity from an int to a double.

+++ src/Herald.h
@@ -47,7 +47,7 @@ enum {
 
 struct InputReq {
 	void* Req;
-	int Quantity;
+	double Quantity;
 };
 
 struct Object {
